ecommerce-backend/requirements.txt
Django==5.0.*
djangorestframework==3.14.*
djangorestframework-simplejwt==5.2.*
django-environ==0.9.*
psycopg2-binary==2.9.*
social-auth-app-django==5.0.*
social-auth-core==4.3.*
dj-stripe==2.8.1
django-storages==1.13.*
boto3==1.23.*
django-ses>=3.5.2
django-hosts==5.2.*
requests==2.28.*
openai>=0.27.2
pydantic>=1.10.7
django-extensions==3.2.*
django-redis==5.4.*
celery>=5.3.6
django-celery-results>=2.5.1
django-celery-beat>=2.6.0
redis>=4.5.0
sentry-sdk==1.36.*
drf-yasg==1.21.*
django-ratelimit>=4.0.0
gunicorn==22.0.0
whitenoise==6.3.*
channels>=4.0.0
channels-redis>=4.2.0
daphne>=4.0.0
